Updating examples with correctly formatted YAML + remove unnecessary files
