+++ examples/subkeys.yaml
@@ -1,10 +1,13 @@
-subkeys:
-    myconf:
-        - a: "do"
-        - b: "re"
-        - c: "mi"
-    servers:
-        useast:
-            - host: "host01"
-        uswest:
-            - host: "host02"
+app-domain:
+  coolapp:
+    coolapp-service:
+     configs:
+       a: "do"
+       b: "re"
+       c: "mi"
+     servers:
+       useast:
+         host: "host01"
+       uswest:
+         host: "host02"
+
